refactor(jangar): split torghut strike parsing
